+++ changelog.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## Version 0.4.0 (2023/07/18)
+
+- Added images in text result
+- Fixed image parsing crash; Closes #1
+
 ## Version 0.3.0 (2023/07/18)
 
 - Added custom exceptions to catch a specific error
-  
+
 ## Version 0.2.1 (2023/07/18)
 
 - Exported `prompt`